[Phase 0.5 / 0.5.4a] header: cleanup dead files, rename navbar-menu, merge title block
Closes #49.

- Move src/components/ui/aceternity/navbar-menu.jsx -> src/components/layout/header/NavMenu.tsx (same content; minimal type casts to compile as .tsx)
- Delete empty src/components/ui/aceternity/ directory
- Create HeaderTitle.tsx merging HeaderTitleBlock + WebLogo content (WebTitle was dead/commented out)
- Delete dead files: Menu.jsx, MobileMenu.jsx, MobileMenuList.jsx, MobileMenuCloseButton.jsx, WebTitle.jsx, HeaderTitleBlock.jsx, WebLogo.jsx, header.module.css, features/jobs-curator/layout/HeaderTitle.tsx
- Header.jsx: update imports to new paths (./NavMenu, ./HeaderTitle); remove isJobsCurator state and JobsCuratorHeaderTitle import
