--- src/components/layout/header/MobileMenuCloseButton.jsx
@@ -1,20 +1,0 @@
-import React from "react";
-import CancelIcon from "@/deprecated-components/ui/CancelIcon";
-
-export default function MobileMenuCloseButton({
-  isMobileMenuOpen,
-  setIsMobileMenuOpen,
-}) {
-  const handleClick = () => {
-    setIsMobileMenuOpen(false);
-  };
-
-  return (
-    <button
-      onClick={handleClick}
-      className="z-20 focus:outline-none focus:bg-transparent"
-    >
-      {isMobileMenuOpen && <CancelIcon />}
-    </button>
-  );
-}
